requirements.txt
GitPython
chardet